Renaming microbloggin to thoughts.
It will not be microblogging. I'm going to write personal thoughts about
various things here (or maybe not). Sometimes I want to write about
things bugging me or an opinion about an article. Gosh, sometimes I even
want to reply to a tweet and have a better place to write my thoughts.

This is the place for it. Let's see.