Add a text about piracy.
I've been discussing this with friends for years. I'm not sure if this
post is exactly what I wanted to say, but I will reserve the rights of
editing it later.

I'm writing this now (after midnight) after taking the day "off" after
being sick, doing a COVID test (negative), taking care of two babies the
whole day, and feeling a bit sleepy. That's the only time I'm finding to
write, so take it with a grain of salt.

PS. it felt like a git commit therapy writing this.