Remove Disqus for now.
It's been a while since I wanted to do this but I was hesitant.

The main problem is: I'm a big fan of comments and discussions. I would
love to have these discussions in place, so other people will have full
context when the find the content. That would be the best scenarion IMO.

In practice: people are not commenting anymore. It's sad, but it's true.
People read the content when you post on social media and add their
comments there. Their comments are lost forever in the sea of
time-series messages.

I was trying to fight this, but... nobody cares. I'm just going to
remove extra JavaScript from my website 🤷.